feat: POST /email
