+++ app/backend/src/controllers/Email.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from 'express';
 import { StatusCodes } from 'http-status-codes';
+import IEmail from '../interfaces/IEmail';
 import EmailService from '../services/Email';
 
 export default class EmailController {
@@ -17,5 +18,12 @@ export default class EmailController {
     await EmailService.updateEmailById(parseInt(id, 10), email);
 
     res.status(StatusCodes.OK).json({ message: 'Email has been updated successfully' });
+  }
+
+  public static async createEmail(req: Request, res: Response) {
+    const { ownerId, email } = req.body as IEmail;
+    const createdEmail = await EmailService.createEmail({ ownerId, email });
+
+    res.status(StatusCodes.CREATED).json(createdEmail);
   }
 }
