+++ app/backend/src/middlewares/Email.ts
@@ -1,5 +1,6 @@
 import { Request, Response, NextFunction } from 'express';
 import { StatusCodes } from 'http-status-codes';
+import IEmail from '../interfaces/IEmail';
 import EmailService from '../services/Email';
 
 export default class EmailMiddleware {
@@ -8,6 +9,19 @@ export default class EmailMiddleware {
     const contact = await EmailService.getEmailByParam(parseInt(id, 10));
 
     if (!contact) return next({ code: StatusCodes.NOT_FOUND, message: 'Email not found' });
+
+    next();
+  }
+
+  public static async validateCreateEmailBody(req: Request, _res: Response, next: NextFunction) {
+    const { email, ownerId } = req.body as IEmail;
+
+    if (!email || !ownerId) {
+      return next({
+        code: StatusCodes.BAD_REQUEST,
+        message: 'email and ownerId values must be provided.',
+      });
+    }
 
     next();
   }
@@ -25,9 +39,8 @@ export default class EmailMiddleware {
 
     const emailInUse = await EmailService.getEmailByParam(email, 'email');
 
-    if (!emailInUse) {
-      next();
-    }
+    if (!emailInUse) return next();
+
     return next({
       code: StatusCodes.CONFLICT,
       message: 'Email is already in use.',
