+++ app/backend/src/routes/Email.ts
@@ -1,5 +1,6 @@
 import express from 'express';
 import EmailController from '../controllers/Email';
+import ContactMiddleware from '../middlewares/Contact';
 import EmailMiddleware from '../middlewares/Email';
 
 const EmailRouter: express.Router = express.Router();
@@ -10,6 +11,13 @@ EmailRouter.patch(
   EmailMiddleware.validateId,
   EmailMiddleware.validateEmail,
   EmailController.updateEmailById,
+);
+EmailRouter.post(
+  '/',
+  EmailMiddleware.validateCreateEmailBody,
+  ContactMiddleware.validateOwnerId,
+  EmailMiddleware.validateEmail,
+  EmailController.createEmail,
 );
 
 export default EmailRouter;
