+++ app/backend/src/services/Email.ts
@@ -1,3 +1,4 @@
+import IEmail from '../interfaces/IEmail';
 import EmailModel from '../models/Email';
 
 type EmailValue = number | string;
@@ -13,5 +14,9 @@ export default class EmailService {
 
   public static async updateEmailById(id: number, email: string) {
     await EmailModel.updateEmailById(id, email);
+  }
+
+  static async createEmail(obj: IEmail) {
+    return EmailModel.createEmail(obj);
   }
 }
